# Fix performance regression in ColumnTransformer