--- a/doc/whats_new/v1.5.rst
+++ b/doc/whats_new/v1.5.rst
@@ -33,6 +33,13 @@ Changes impacting many modules
 Changelog
 ---------
 
+:mod:`sklearn.compose`
+......................
+
+- |Efficiency| Fix a performance regression in :class:`compose.ColumnTransformer`
+  where the full input data was copied for each transformer when `n_jobs > 1`.
+  :pr:`` by :user:`Jérémie du Boisberranger <jeremiedbb>`.
+
 :mod:`sklearn.metrics`
 ......................