--- a/doc/whats_new/v1.5.rst
+++ b/doc/whats_new/v1.5.rst
@@ -38,7 +38,7 @@ Changelog
 
 - |Efficiency| Fix a performance regression in :class:`compose.ColumnTransformer`
   where the full input data was copied for each transformer when `n_jobs > 1`.
-  :pr:`` by :user:`Jérémie du Boisberranger <jeremiedbb>`.
+  :pr:`29330` by :user:`Jérémie du Boisberranger <jeremiedbb>`.
 
 :mod:`sklearn.metrics`
 ......................