Fixes #29229 

#28822 was a quick fix for https://github.com/scikit-learn/scikit-learn/issues/28781 before releasing 1.5.0. It turns out that it caused a performance regression when the input can't be memory mapped, for instance columns where elements are arrays of variable length. In that case the whole `X` is copied in each subprocess.

Since https://github.com/scikit-learn/scikit-learn/pull/29018 has now been merged, we can revert #28822 because it also fixes #28781 without causing the performance regression (note that I kept the non regression test). The following results of the snippet from #29229 with this PR are close to the 1.4.2 ones:
```
sklearn version: 1.6.dev0 and n_jobs=1
5: Per col: 0.0206s / total 0.10 s
10: Per col: 0.0203s / total 0.20 s
15: Per col: 0.0204s / total 0.31 s
20: Per col: 0.0201s / total 0.40 s
25: Per col: 0.0214s / total 0.54 s
30: Per col: 0.0204s / total 0.61 s
35: Per col: 0.0202s / total 0.71 s
40: Per col: 0.0200s / total 0.80 s
45: Per col: 0.0208s / total 0.93 s

sklearn version: 1.6.dev0 and n_jobs=2
5: Per col: 0.1613s / total 0.81 s
10: Per col: 0.0352s / total 0.35 s
15: Per col: 0.0321s / total 0.48 s
20: Per col: 0.0334s / total 0.67 s
25: Per col: 0.0299s / total 0.75 s
30: Per col: 0.0296s / total 0.89 s
35: Per col: 0.0283s / total 0.99 s
40: Per col: 0.0294s / total 1.18 s
45: Per col: 0.0294s / total 1.32 s
```